Report script diagnostics to the host
